test(logic): split init_game_orchestrator tests into test/setup (Refs #2288)
Extract shared locked-full-init AC helpers to init_game_orchestrator_test_support.dart.
Divide runInitGame scenarios into part1/part2 files under test/setup mirroring lib/src/setup.
Removes ~764-line root test file; each part stays within the logic test line budget.
